+++ src/utils/api/api.ts
@@ -1,23 +1,26 @@
+//
 // A super simple (not finished) wrapper for NXT API calls
+//
 
-export interface IAPIResult {
-  notsure: string;
-}
+import { BASEURL } from "./constants";
 
-export async function API<IAPIResult>(
+export async function API(
   url: string,
   method: "GET" | "POST",
   /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
   data?: any
-) {
-  const result = await fetch(url, {
-    headers: {
-      "Content-Type": "application/json",
-      "Access-Control-Allow-Origin": "*",
-    },
-    method,
-    body: JSON.stringify(data),
-  });
+  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+): Promise<any> {
+  let result;
+  if (method === "GET") {
+    result = await fetch(BASEURL + url);
+  } else {
+    result = await fetch(BASEURL + url, {
+      method,
+      body: JSON.stringify(data),
+    });
+  }
+
   const jsonResult = await result.json();
   return jsonResult;
 }