+++ src/utils/api/constants.ts
@@ -1,1 +1,2 @@
-export const BASEURL = "http://nodes.jup.io/nxt?";
+// TODO: Make this configurable based on prod/dev run status
+export const BASEURL = "http://localhost:3080/nxt?"; // proxy runs on 3080 by default