+++ src/utils/api/getAccount.ts
@@ -1,18 +1,24 @@
+//
+// API call helper for getAccount, not meant to be called directly (meant to be used inside the APIProvider)
+//
+
 import { API } from "./api";
-import { BASEURL } from "./constants";
 
 /*
  * http://localhost:7876/nxt?
  *   requestType=getAccount&
- *   account=NXT-4VNQ-RWZC-4WWQ-GVM8S
+ *   account=JUP-9J5L-9BX3-7HCX-AP3MK
  */
 
-function getAccount(account: string) {
+async function getAccount(account: string) {
+  let result;
   try {
-    API(`${BASEURL}requestType=getAccount&account=${account}`, "GET");
+    result = await API(`requestType=getAccount&account=${account}`, "GET");
   } catch (e) {
     console.error("error getAccount():", e);
+    return false;
   }
+  return result;
 }
 
 export default getAccount;