+++ src/utils/api/sendJUP.ts
@@ -1,61 +1,13 @@
-import { IUnsignedTransaction } from "views/Dashboard/Dashboard";
+//
+// API call helper for sendJUP, not meant to be called directly (meant to be used inside the APIProvider)
+//
+
+import { IUnsignedTransaction, ISignedTransaction, IBroadcastTransactionResult, ISignedTransactionResult } from "types/NXTAPI";
 import { API } from "./api";
-import { BASEURL } from "./constants";
 
-export interface ISignedTransaction extends IUnsignedTransaction {
-  signature: string;
-}
-
-// TODO: Implement signing (locally if possible)
+// TODO: Improve validation
+//        - Should validation run before signing in order to report input flaws to the user? probably
 // TODO: Implement broadcasting
-// TODO: Implement validation
-
-// sign
-//
-// http://localhost:7876/nxt?
-//   requestType=signTransaction&
-//   unsignedTransactionBytes=00100cfb3c03a00510f09c34f225d425306e5be55a494690...&
-//   secretPhrase=SecretPhrase
-
-// example unsigned tx, redacted
-//
-// {"senderPublicKey":"abc",
-// "feeNQT":"5000",
-// "type":0,
-// "version":1,
-// "phased":false,
-// "ecBlockId":"123",
-// "attachment":{"version.OrdinaryPayment":0},
-// "senderRS":"JUP-XXXX-XXXX-XXXX-XXXXX",
-// "subtype":0,"amountNQT":"100000000000",
-// "sender":"123",
-// "recipientRS":"JUP-XXXX-XXXX-XXXX-XXXXX",
-// "recipient":"123",
-// "ecBlockHeight":0,
-// "deadline":1440,
-// "timestamp":141421180,
-// "height":2147483647}
-
-// signed tx example (redacted)
-//
-// {"senderPublicKey":"123",
-// "feeNQT":"5000",
-// "type":0,
-// "version":1,
-// "phased":false,
-// "ecBlockId":"123",
-// "attachment":{"version.OrdinaryPayment":0},
-// "senderRS":"JUP-ABCD-ABCD-ABCD-ABCDE",
-// "subtype":0,
-// "amountNQT":"100000000000",
-// "sender":"123",
-// "recipientRS":"JUP-ABCD-ABCD-ABCD-ABCDE",
-// "recipient":"123",
-// "ecBlockHeight":0,
-// "deadline":1440,
-// "timestamp":141423231,
-// "height":2147483647,
-// "signature":"abc123"}
 
 // broadcast
 //
@@ -64,24 +16,24 @@ export interface ISignedTransaction extends IUnsignedTransaction {
 //   transactionBytes=001046aac6013c0057fb6f3a958e320bb49c4e81b4c2cf28b9f25d086c143
 
 // sendTransaction call to the API requires adminPassword so that cannot be used
-function sendJUP(unsigned: IUnsignedTransaction) {
-  let signedTx: ISignedTransaction | undefined;
+async function sendJUP(unsigned: IUnsignedTransaction) {
+  let signedTx: ISignedTransaction;
   let isValid: boolean;
   try {
     // sign
-    console.log("signing tx...", unsigned);
-    signedTx = signTx(unsigned);
+    signedTx = await signTx(unsigned);
     // validate
     isValid = validateTx(signedTx);
-    // send
+    console.log("isValid:", isValid);
+    // // broadcast
     if (isValid) {
-      console.log("valid transaction, broadcasting not implemented yet. signedTx:", signedTx, "isValid:", isValid);
-      return;
+      const broadcastResult = await broadcastTx(signedTx);
+      broadcastResult ? true : false;
     }
-    console.error("transaction invalid");
-    return;
+    return false;
   } catch (e) {
     console.error("error sendJUP():", e);
+    return false;
   }
 }
 
@@ -89,22 +41,32 @@ function sendJUP(unsigned: IUnsignedTransaction) {
 // Helper functions
 //
 
-function signTx(unsigned: IUnsignedTransaction) {
+async function signTx(unsigned: IUnsignedTransaction) {
   const secret = "test"; // TODO: implement
 
-  let result;
+  console.log("preparing to sign JSON:", unsigned);
+
+  let result: ISignedTransactionResult;
   try {
-    // result = API(BASEURL + "requestType=signTransaction&unsignedTransactionBytes=" + unsigned + "&secretPhrase=" + secret, "GET");
+    result = await API("requestType=signTransaction&unsignedTransactionJSON=" + JSON.stringify(unsigned) + "&secretPhrase=" + secret, "GET");
+    if (result?.transactionJSON?.signature) {
+      return { ...unsigned, signature: result.transactionJSON.signature }; // signing was a success, return the new object
+    }
+
+    console.log("got result from sign:", result);
   } catch (e) {
     console.error("error while signing tx:", e);
-    return;
+    return { ...unsigned, signature: "" };
   }
-  return { ...unsigned, signature: "test" }; // TODO: implement, just passes in a test signature for now
+  return { ...unsigned, signature: "" };
 }
 
 function validateTx(signed?: ISignedTransaction) {
   if (signed === undefined) {
     return false;
+  } else if (signed.signature == "") {
+    console.log("no signature in transaction to be validated:", signed);
+    return false; // might want to handle this differently
   }
   console.log("validating tx:", signed, "with signature:", signed.signature);
 
@@ -113,7 +75,25 @@ function validateTx(signed?: ISignedTransaction) {
     return false;
   }
 
-  return true; // TODO: implement
+  return true;
+}
+
+async function broadcastTx(signed: ISignedTransaction): Promise<false | IBroadcastTransactionResult> {
+  let result: IBroadcastTransactionResult;
+  try {
+    result = await API("requestType=broadcastTransaction&transactionJSON=" + JSON.stringify(signed), "POST");
+
+    // make sure the broadcast succeeded
+    if (result.transaction) {
+      console.log("broadcast tx:", result.transaction, "with hash:", result.fullHash);
+      return result;
+    }
+    console.log("broadcast resulted in unexpected result, investigate this:", result);
+  } catch (e) {
+    console.error("error while signing tx:", e);
+    return false;
+  }
+  return false;
 }
 
 export default sendJUP;